refactor(ui): fix code quality issues in metrics implementation
- Extract formatLastUpdated function to shared utility in lib/utils.ts
- Remove unused MetricsControlBar component that was never imported
- Fix potential memory leak by using useRef to prevent repeated initial refreshes
- Improve component maintainability and reduce duplication
